requirements.txt
streamlit
folium
streamlit-folium
llama-index
llama-index-llms-openrouter
requests
python-dotenv